refactor: 패키지 경로 수정
admininquiry 에서 directinquiry로 변경
