--- src/main/java/com/biddingmate/biddinggo/admininquiry/service/AdminInquiryService.java
@@ -1,17 +1,0 @@
-package com.biddingmate.biddinggo.admininquiry.service;
-
-import com.biddingmate.biddinggo.admininquiry.dto.AdminInquiryView;
-import com.biddingmate.biddinggo.admininquiry.dto.AdminInquiryViewDetail;
-import com.biddingmate.biddinggo.admininquiry.dto.AnswerAdminInquiryRequest;
-import com.biddingmate.biddinggo.admininquiry.dto.AnswerAdminInquiryResponse;
-import com.biddingmate.biddinggo.admininquiry.dto.CreateAdminInquiryRequest;
-import com.biddingmate.biddinggo.admininquiry.dto.CreateAdminInquiryResponse;
-import com.biddingmate.biddinggo.common.request.BasePageRequest;
-import com.biddingmate.biddinggo.common.response.PageResponse;
-
-public interface AdminInquiryService {
-    CreateAdminInquiryResponse createAdminInquiry(CreateAdminInquiryRequest request);
-    PageResponse<AdminInquiryView> findAdminInquiry(BasePageRequest request, boolean isAdmin, Long memberId);
-    AdminInquiryViewDetail findAdminInquiryDetail(Long inquiryId, boolean isAdmin, Long memberId);
-    AnswerAdminInquiryResponse answerAdminInquiry(Long inquiryId, AnswerAdminInquiryRequest request, Long adminId);
-}
