+++ src/main/java/com/biddingmate/biddinggo/directinquiry/controller/DirectInquiryController.java
@@ -1,12 +1,12 @@
-package com.biddingmate.biddinggo.admininquiry.controller;
+package com.biddingmate.biddinggo.directinquiry.controller;
 
-import com.biddingmate.biddinggo.admininquiry.dto.AdminInquiryView;
-import com.biddingmate.biddinggo.admininquiry.dto.AdminInquiryViewDetail;
-import com.biddingmate.biddinggo.admininquiry.dto.AnswerAdminInquiryRequest;
-import com.biddingmate.biddinggo.admininquiry.dto.AnswerAdminInquiryResponse;
-import com.biddingmate.biddinggo.admininquiry.dto.CreateAdminInquiryRequest;
-import com.biddingmate.biddinggo.admininquiry.dto.CreateAdminInquiryResponse;
-import com.biddingmate.biddinggo.admininquiry.service.AdminInquiryService;
+import com.biddingmate.biddinggo.directinquiry.dto.AdminInquiryView;
+import com.biddingmate.biddinggo.directinquiry.dto.AdminInquiryViewDetail;
+import com.biddingmate.biddinggo.directinquiry.dto.AnswerAdminInquiryRequest;
+import com.biddingmate.biddinggo.directinquiry.dto.AnswerAdminInquiryResponse;
+import com.biddingmate.biddinggo.directinquiry.dto.CreateAdminInquiryRequest;
+import com.biddingmate.biddinggo.directinquiry.dto.CreateAdminInquiryResponse;
+import com.biddingmate.biddinggo.directinquiry.service.DirectInquiryService;
 import com.biddingmate.biddinggo.common.request.BasePageRequest;
 import com.biddingmate.biddinggo.common.response.ApiResponse;
 import com.biddingmate.biddinggo.common.response.PageResponse;
@@ -25,9 +25,9 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/api/v1/admin-inquiries")
-public class AdminInquiryController {
-    private final AdminInquiryService adminInquiryService;
+@RequestMapping("/api/v1/direct-inquiries")
+public class DirectInquiryController {
+    private final DirectInquiryService directInquiryService;
 
     @PostMapping("")
     public ResponseEntity<ApiResponse<CreateAdminInquiryResponse>> createAdminInquiry(
@@ -35,7 +35,7 @@ public class AdminInquiryController {
     ) {
         // 인증 인가 전이므로 writerId를 하드코딩했습니다.
         // 추후에 리펙터링 대상입니다.
-        CreateAdminInquiryResponse result = adminInquiryService.createAdminInquiry(request);
+        CreateAdminInquiryResponse result = directInquiryService.createAdminInquiry(request);
 
         return ApiResponse.of(HttpStatus.OK, null, "1대1 문의 성공", result);
     }
@@ -46,7 +46,7 @@ public class AdminInquiryController {
                                                                                         @RequestParam long memberId) {
         // 인증 인가 완료 전이므로 role 검사 및 memberId 전송
         // 추후에 리펙터링 대상입니다.
-        PageResponse<AdminInquiryView> result = adminInquiryService.findAdminInquiry(request, isAdmin, memberId);
+        PageResponse<AdminInquiryView> result = directInquiryService.findAdminInquiry(request, isAdmin, memberId);
 
         return ApiResponse.of(HttpStatus.OK, null, "1대1 문의 목록 조회 성공", result);
     }
@@ -55,7 +55,7 @@ public class AdminInquiryController {
     public ResponseEntity<ApiResponse<AdminInquiryViewDetail>> findAdminInquiryDetail(@PathVariable Long inquiryId,
                                                                                       @RequestParam boolean isAdmin,
                                                                                       @RequestParam Long memberId) {
-        AdminInquiryViewDetail result = adminInquiryService.findAdminInquiryDetail(inquiryId, isAdmin, memberId);
+        AdminInquiryViewDetail result = directInquiryService.findAdminInquiryDetail(inquiryId, isAdmin, memberId);
 
         return ApiResponse.of(HttpStatus.OK, null, "1대1 문의 상세 조회 성공", result);
     }
@@ -66,7 +66,7 @@ public class AdminInquiryController {
     public ResponseEntity<ApiResponse<AnswerAdminInquiryResponse>> answerAdminInquiry(@PathVariable Long inquiryId,
                                                                                      @Valid @RequestBody AnswerAdminInquiryRequest request,
                                                                                       @RequestParam Long adminId) {
-        AnswerAdminInquiryResponse result = adminInquiryService.answerAdminInquiry(inquiryId, request, adminId);
+        AnswerAdminInquiryResponse result = directInquiryService.answerAdminInquiry(inquiryId, request, adminId);
 
         return ApiResponse.of(HttpStatus.OK, null, "1대1 문의 답변 등록 성공", result);
     }
