+++ src/main/resources/mapper/admin-inquiry/admin-inquiry-mapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.biddingmate.biddinggo.admininquiry.mapper.AdminInquiryMapper">
+<mapper namespace="com.biddingmate.biddinggo.directinquiry.mapper.AdminInquiryMapper">
     <insert id="insert" parameterType="AdminInquiry" keyProperty="id" useGeneratedKeys="true" >
         INSERT INTO admin_inquiry (
             writer_id,
